Move enable/disable deeplinking methods to utils
